+++ src/cryptography/hazmat/primitives/serialization/pkcs7.py
@@ -21,6 +21,7 @@ from cryptography.hazmat.primitives.ciphers import (
     algorithms,
 )
 from cryptography.utils import _check_byteslike
+from cryptography.x509.verification import PolicyBuilder, Store
 
 load_pem_pkcs7_certificates = rust_pkcs7.load_pem_pkcs7_certificates
 
@@ -51,6 +52,8 @@ class PKCS7Options(utils.Enum):
     NoCapabilities = "Don't embed SMIME capabilities"
     NoAttributes = "Don't embed authenticatedAttributes"
     NoCerts = "Don't embed signer certificate"
+    NoVerify = "Don't verify signers certificate"
+    NoSigs = "Don't verify signature"
 
 
 class PKCS7SignatureBuilder:
@@ -184,6 +187,11 @@ class PKCS7SignatureBuilder:
             )
 
         return rust_pkcs7.sign_and_serialize(self, encoding, options)
+
+
+pkcs7_verify_der = rust_pkcs7.verify_der
+pkcs7_verify_pem = rust_pkcs7.verify_pem
+pkcs7_verify_smime = rust_pkcs7.verify_smime
 
 
 class PKCS7EnvelopeBuilder:
@@ -356,6 +364,39 @@ def _smime_signed_encode(
     )
     g.flatten(m)
     return fp.getvalue()
+
+
+def _smime_signed_decode(data: bytes) -> tuple[bytes | None, bytes]:
+    message = email.message_from_bytes(data)
+    content_type = message.get_content_type()
+    if content_type == "multipart/signed":
+        payload = message.get_payload()
+        if not isinstance(payload, list):
+            raise ValueError(
+                "Malformed multipart/signed message: must be multipart"
+            )
+        assert isinstance(payload[0], email.message.Message), (
+            "Malformed multipart/signed message: first part (content) "
+            "must be a MIME message"
+        )
+        assert isinstance(payload[1], email.message.Message), (
+            "Malformed multipart/signed message: second part (signature) "
+            "must be a MIME message"
+        )
+        return (
+            bytes(payload[0].get_payload(decode=True)),
+            bytes(payload[1].get_payload(decode=True)),
+        )
+    elif content_type == "application/pkcs7-mime":
+        return None, bytes(message.get_payload(decode=True))
+    else:
+        raise ValueError("Not an S/MIME signed message")
+
+
+def _verify_pkcs7_certificates(certificates: list[x509.Certificate]) -> None:
+    builder = PolicyBuilder().store(Store(certificates))
+    verifier = builder.build_client_verifier()
+    verifier.verify(certificates[0], certificates[1:])
 
 
 def _smime_enveloped_encode(data: bytes) -> bytes: