+++ CHANGELOG.rst
@@ -12,6 +12,10 @@ Changelog
   ``cryptography`` release.
 * Added support for PKCS7 decryption & encryption using AES-256 as content algorithm, 
   in addition to AES-128. 
+* Added basic support for PKCS7 verification (including S/MIME 3.2) via
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_verify_der`,
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_verify_pem`, and
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_verify_smime`.
 
 .. _v44-0-0: